+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add a downloadable ICS file to scheduled student classrooms
 - frontend bulk delete for videos, webinars and classrooms
 - Can choose inactive resource by site (#2276)
+- Setting for instructor classroom invitation link expiration 
+- ClassroomRecording Delete API
+  - delete BBB recordings
+  - delete ClassroomsRecording
 
 ### Changed